% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns.R
\name{as_unit_vector}
\alias{as_unit_vector}
\title{transform unbounded to unit vector}
\usage{
as_unit_vector(y)
}
\arguments{
\item{y}{numeric vector}
}
\description{
transform unbounded to unit vector
}
\examples{
x = as_unit_vector(rnorm(5))
(sqrt(x \%*\% x) == 1)
}
