% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns.R
\name{make_iid_matrix}
\alias{make_iid_matrix}
\title{make design matrix for IID random effect}
\usage{
make_iid_matrix(x)
}
\arguments{
\item{x}{subject ID column}
}
\description{
make design matrix for IID random effect
}
