% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns.R
\name{simplex_transform}
\alias{simplex_transform}
\title{transform an unbounded (normally distributed) to simplex}
\usage{
simplex_transform(y)
}
\arguments{
\item{y}{a vector of lenth J - 1}
}
\value{
an simplex of length J
}
\description{
transform an unbounded (normally distributed) to simplex
}
\examples{
sum(simplex_transform(rnorm(5)))
}
